Fix rv64 build issue.

+++ src/kernels/cpu/optimized/slice.cpp
@@ -38,7 +38,7 @@ void _slice_contiguous_dim_copy(const runtime_shape_t &begins, NNCASE_UNUSED con
     Callable &&line_copy, runtime_shape_t &in_index,
     std::true_type) noexcept
 {
-    for (size_t i = begins[CurDim]; i < ends[CurDim]; ++i)
+    for (size_t i = begins[CurDim]; i < static_cast<size_t>(ends[CurDim]); ++i)
     {
         in_index[CurDim] = i;
         _slice_contiguous_dim_copy<Dims, CurDim + 1>(begins, ends,
@@ -58,7 +58,7 @@ result<void> slice_contiguous_impl(const T *input, T *output, const runtime_shap
     runtime_shape_t in_index(in_shape.size());
 
     auto line_copy = [&]() {
-        const auto distance = ends[dims] - begins[dims];
+        const auto distance = static_cast<size_t>(ends[dims]) - begins[dims];
         const auto copy_size = distance * elemsize;
         const auto *in_ptr = input + offset(in_strides, in_index);
         memcpy(out_ptr, in_ptr, copy_size);
@@ -106,7 +106,7 @@ void _slice_dim_copy(const runtime_shape_t &begins, const runtime_axis_t &ends,
     std::true_type) noexcept
 {
     out_index[CurDim] = 0;
-    for (size_t i = begins[CurDim]; i < ends[CurDim]; i += strides[CurDim])
+    for (size_t i = begins[CurDim]; i < static_cast<size_t>(ends[CurDim]); i += strides[CurDim])
     {
         in_index[CurDim] = i;
         _slice_dim_copy<Dims, CurDim + 1>(begins, ends, strides,
@@ -160,7 +160,7 @@ result<void> slice_linecopy_impl(const T *input, T *output, const runtime_shape_
         in_shape, begins, ends, strides,
         [&, dims](runtime_shape_t &in_index, runtime_shape_t &out_index) {
             in_index[dims] = begins[dims];
-            const auto distance = ends[dims] - begins[dims];
+            const auto distance = static_cast<size_t>(ends[dims]) - begins[dims];
             auto copy_size = distance * sizeof(T);
             const auto *in_ptr = input + offset(in_strides, in_index);
             auto *out_ptr = output + offset(out_strides, out_index);
@@ -176,7 +176,7 @@ result<void> slice_strides_impl(const T *input, T *output, const runtime_shape_t
     return _slice_impl(
         in_shape, begins, ends, strides,
         [&, dims](runtime_shape_t &in_index, runtime_shape_t &out_index) {
-            for (size_t i = begins[dims]; i < ends[dims]; i += strides[dims])
+            for (size_t i = begins[dims]; i < static_cast<size_t>(ends[dims]); i += strides[dims])
             {
                 in_index[dims] = i;
                 output[offset(out_strides, out_index)] = input[offset(in_strides, in_index)];
@@ -207,7 +207,7 @@ result<void> optimized::slice(datatype_t type, const gsl::byte *input, gsl::byte
     runtime_shape_t out_shape(dims);
     for (size_t i = 0; i < dims; ++i)
     {
-        out_shape[i] = ends[i] - begins[i];
+        out_shape[i] = static_cast<size_t>(ends[i]) - begins[i];
     }
 
     for (size_t i = 0; i < dims; ++i)
